Refactor webpack config out to support/webpack.config.js
