+++ gulpfile.js
@@ -14,34 +14,13 @@ gulp.task('default', ['build']);
 // BUILDING
 ////////////////////////////////////////
 
-const BUILD_TARGET_FILENAME = 'socket.io.js';
 const BUILD_TARGET_DIR = './';
 
 gulp.task('build', ['webpack']);
 
 gulp.task('webpack', function() {
   return gulp.src('lib/*.js')
-    .pipe(webpack({
-      entry: './lib/index.js',
-      output: {
-        library: 'io',
-        libraryTarget: 'umd',
-        filename: BUILD_TARGET_FILENAME
-      },
-      externals: {
-          'global': glob()
-      },
-      module: {
-        loaders: [{
-          test: /\.(js|jsx)?$/,
-          exclude: /(node_modules|bower_components)/,
-          loader: 'babel', // 'babel-loader' is also a legal name to reference
-          query: {
-              presets: ['react', 'es2015']
-          }
-        }]
-      }
-    }))
+    .pipe(webpack(require('./support/webpack.config.js')))
     .pipe(gulp.dest(BUILD_TARGET_DIR));
 });
 
@@ -124,15 +103,3 @@ gulp.task('test-cov', ['istanbul-pre-test'], function(){
       process.exit();
     });
 });
-
-/**
- * Populates `global`.
- *
- * @api private
- */
-
-function glob(){
-  return 'typeof self !== "undefined" ? self : '
-    + 'typeof window !== "undefined" ? window : '
-    + 'typeof global !== "undefined" ? global : {}';
-}
